backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.24.0
opencv-python>=4.8.1
mediapipe==0.10.8
numpy>=1.24.3,<2.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyJWT==2.8.0
pydantic==2.10.5
pydantic-settings==2.7.1
python-dotenv==1.0.0
email-validator==2.2.0
